Changes get_oci_metadata signature
It now includes an additional parameter "arch" which is used to search
only artifacts that match that architecture. before the architectures
"amd64" and "arm64" where assumed. This leads to a file not found error
if the build directory only contains artifacts for one architecture.
Also if artifacts are found they are pushed which is also not desired
behaviour when calling a program with --arch <arch>

Signed-off-by: Malte Münch <muench@b1-systems.de>